// src/@types/actions.d.ts
export interface Action{
    payload: uknown
    type: string
}


/**
 * WAllet
 * Address
 * ChatWithChatbot
 * Online Consultation later
 * Phamacy
 * Indivdual
 * Inventory
 * Authentication.
 * Blockchanin: coin
 * Token Discounts 
 * NFTs/Tokens - 
 * Convert
 * 
 * Phamacies
 * - Name
 * - Role based
 * 
 * Search for drugs and where to find them
 * Get tokens by watching videos.
 */